Consolidate building MeasureSpecs from Metric objects into helper
There is a helper method in MetricSemantics that takes a Measure
model object and extracts the MeasureSpecs required for computing
that metric.

For some reason we have an exact duplicate of that logic inlined
into the core of the DataflowPlanBuilder. Since we will need to
modify this logic, and wish for it to be consistent across both
contexts, this commit replaces the duplicate code with a call
to the helper method.