Move where constraint conversion logic out of the query parser
For one reason or another the MetricFlowQueryParser contained logic
for doing conversions from WhereClauseConstraint instances to
SpecWhereClauseConstraint instances. This particular set of
transformations makes the most sense as either a static factory
on SpecWhereClauseConstraint or a method on WhereClauseConstraint.

Up until now everything was fine because we only ever did that
conversion in the query parser itself, but now that we are allowing
WhereClauseConstraints on measure inputs to metrics we need to
broaden that scope of access, which then results in circular
import errors.

Moving these methods to the more logically appropriate places still
produces circular import dependencies, but we can resolve them with
module-level imports. Keeping these methods on the WhereClauseConstraint
makes the most sense from a class organization perspective. A
larger module refactor to produce a more permanent fix for the
circular imports will have to wait for another day, as it's likely to
be quite a large copy/paste kind of change.