Create and use a custom jsonschema validator based on Draft7Validator
We want to be able to ignore all properties in config documents that
match the regex '^__(.*)__$', e.g. '__parsing_context__',
'__custom_property_name__', and etc. To do this without actually
adding a patternProperty to every schema, we had to override the
'additonalProperties' validator of the Draft7Validator. As we can
not actually change out the validators used by the Draft7Validator,
we had to write our own custom validator, which we call 'SchemaValidator'